Extract wms layer props utils to map-utils
